Restructure API services; Update for user-roles
