+++ src/pages/Login/Login.js
@@ -5,7 +5,7 @@ import styles from "./styles";
 import { withStyles } from "@material-ui/core/styles";
 import { withRouter } from "react-router-dom";
 import { UserContext } from "../../services/UserContext";
-import firebase from "../../services/firebase";
+
 import {
   Paper,
   Avatar,
@@ -26,9 +26,8 @@ const Login = props => {
 
   const authenticate = async () => {
     try {
-      const { user } = await firebase.login(email, password);
+      const { user } = await login(email, password); // TODO:
       auth.setUser(user);
-      localStorage.setItem("user", JSON.stringify(user));
       history.push("/");
     } catch (error) {
       alert(error.message);
