+++ src/services/API/index.js
@@ -1,0 +1,2 @@
+export * from "./Groups";
+export * from "./Students";
